+++ src/host/main.cpp
@@ -9,6 +9,7 @@
 #include "ds/non_blocking.h"
 #include "ds/oversized.h"
 #include "enclave.h"
+#include "enclave/reconfiguration_type.h"
 #include "handle_ring_buffer.h"
 #include "load_monitor.h"
 #include "node_connections.h"
@@ -482,6 +483,18 @@ int main(int argc, char** argv)
     ->check(CLI::PositiveNumber)
     ->type_name("UINT");
 
+  ReconfigurationType reconfiguration_type =
+    ReconfigurationType::ONE_TRANSACTION;
+  std::vector<std::pair<std::string, ReconfigurationType>>
+    reconfiguration_type_map{
+      {"1tx", ReconfigurationType::ONE_TRANSACTION},
+      {"2tx", ReconfigurationType::TWO_TRANSACTION}};
+  start
+    ->add_option(
+      "--reconfiguration-type", reconfiguration_type, "Reconfiguration type")
+    ->transform(
+      CLI::CheckedTransformer(reconfiguration_type_map, CLI::ignore_case));
+
   auto join = app.add_subcommand("join", "Join existing network");
   join->configurable();
 
@@ -829,6 +842,7 @@ int main(int argc, char** argv)
 
     ccf_config.startup_host_time = crypto::OpenSSL::to_x509_time_string(
       std::chrono::system_clock::to_time_t(startup_host_time));
+    ccf_config.genesis.reconfiguration_type = reconfiguration_type;
 
     if (*start)
     {